refactor(net): dedup ja4 hash12 helper into shared module
